Restructured finder
